Reworked movie patcher and API

+++ include/GFxArray.h
@@ -32,17 +32,16 @@ public:
 		return RE::GFxValue::SetElement(a_index, &a_value);
 	}
 
-	std::uint32_t FindElement(const RE::GFxValue& a_value)
+	std::int32_t FindElement(const RE::GFxValue& a_value)
 	{
-
 		for (std::uint32_t i = 0; i < GetArraySize(); i++) 
 		{
-			if (GetElement(i) == a_value) 
+			if (GetElement(i) == a_value)
 			{
 				return i;
 			}
 		}
 
-		return static_cast<std::uint32_t>(-1);
+		return -1;
 	}
 };
